silent warning.

+++ src/buildin.c
@@ -26,7 +26,7 @@ builtin(struct AST *ast)
 	if ((ast->type != NODE_COMMAND && ast->type != NODE_AMPR) ||
 			!ast->token)
 		return 0;
-	for (int i = 0; i < LEN(builtins); i++) {
+	for (unsigned long int i = 0; i < LEN(builtins); i++) {
 		if (!strcmp(builtins[i].name, ast->token[0])) {
 			(*builtins[i].func)(ast->token);
 			return 1;
